move stateful modules into state/
