+++ frontend/src/state/response-handler-binding.ts
@@ -1,4 +1,4 @@
 // This file is instantiated by wasm-bindgen in `/frontend/wasm/src/lib.rs` and re-exports the `handleResponse` function to
 // provide access to the global copy of `response-handler.ts` with its shared state, not an isolated duplicate with empty state
 
-export { handleResponse } from "@/utilities/response-handler";
+export { handleResponse } from "@/state/response-handler";
